style: 리뷰 반영 뷰 코드 분리 및 서브뷰 파일 분리
